additional project root folder check

+++ build/docker-rs.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
+if [[ ! -e Cargo.toml || ! -e .git ]]; then
+    printf "not in project root\n"
+    exit -1
+fi
+
 docker run --rm rs-chdiff:latest
